requirements.txt
anthropic>=0.40.0
openai>=1.0.0
google-genai>=0.8.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
scipy>=1.11.0
statsmodels>=0.14.0
